Refactored message handling to improve message archiving and error handling
