+++ grails-app/services/org/pih/warehouse/inventory/StockMovementService.groovy
@@ -188,29 +188,6 @@ class StockMovementService {
         }
     }
 
-    void acceptStockMovement(StockMovement stockMovement) {
-        Shipment shipment = stockMovement?.shipment
-        EventType eventType = EventType.findByEventCode(EventTypeCode.ACCEPTED)
-        Event event = new Event(eventType: eventType, eventDate: new Date())
-        shipment.addToEvents(event)
-        shipment.save(flush:true)
-        notificationService.sendShipmentAcceptedNotification(shipment, shipment.origin, [RoleType.ROLE_SHIPMENT_NOTIFICATION])
-    }
-
-    void attachDocument(StockMovement stockMovement, String fileName, String fileContents) {
-        Document document = new Document()
-        document.documentType = DocumentType.get(Constants.DEFAULT_DOCUMENT_TYPE_ID)
-        document.name = fileName
-        document.filename = fileName
-        document.fileContents = fileContents.bytes
-
-        // FIXME we need to figure out a way to detect the mimetype of the file
-        document.contentType = "application/octet-stream"
-
-        stockMovement.shipment.addToDocuments(document)
-        stockMovement.shipment.save(flush:true)
-    }
-
     void updateRequisitionStatus(String id, RequisitionStatus status) {
 
         log.info "Update status ${id} " + status
@@ -230,7 +207,6 @@ class StockMovementService {
             requisition.save(flush: true)
         }
     }
-
 
     StockMovement updateStockMovement(StockMovement stockMovement) {
         if (!stockMovement.validate()) {
